customExceptions

+++ src/main/java/com/stackroute/muzix/service/MuzixServiceimpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.muzix.service;
 
 import com.stackroute.muzix.domain.Track;
+import com.stackroute.muzix.exceptions.TrackAlreadyExistsException;
+import com.stackroute.muzix.exceptions.TrackNotFoundException;
 import com.stackroute.muzix.repository.MuzixRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,7 +13,7 @@ import java.util.Optional;
 
 @Service
 public class MuzixServiceimpl implements MuzixService {
-private MuzixRepository muzixRepository;
+    private MuzixRepository muzixRepository;
 
     public MuzixServiceimpl() {
     }
@@ -22,46 +24,54 @@ private MuzixRepository muzixRepository;
     }
 
     @Override
-    //saves the track
-    public Track saveTrack(Track track) {
-Track saveedTrack=muzixRepository.save(track);
-return saveedTrack;
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException
+    {
+        if(muzixRepository.existsById(track.getTrackid())){
+            throw new TrackAlreadyExistsException("id already exists");
+        }
+        Track saveedTrack=muzixRepository.save(track);
+        return track;
     }
 
     @Override
-    //delete the track based on id
-    public int deleteTrack(int trackId) {
-    muzixRepository.deleteById(trackId);
+    public int deleteTrack(int trackId) throws TrackNotFoundException {
+        Track track=new Track();
+        if(!muzixRepository.findById(trackId).isPresent()){
+            throw new TrackNotFoundException("id not found");
+        }
+        muzixRepository.deleteById(trackId);
         return trackId;
     }
 
     @Override
-    //retrive all tracks
     public List<Track> getAllTracks() {
-    return muzixRepository.findAll();
+        return muzixRepository.findAll();
     }
 
     @Override
-    //retrive track based on id
-    public Optional<Track> getTrackById(int trackId) {
-    return muzixRepository.findById(trackId);
+    public Optional<Track> getTrackById(int trackId) throws TrackNotFoundException {
+        if(!muzixRepository.findById(trackId).isPresent()){
+            throw new TrackNotFoundException("id not found");
+        }
+        return muzixRepository.findById(trackId);
     }
 
     @Override
-    //update track based on id
-    public Track updateById(Track track, int trackId) {
+    public Track updateById(Track track, int trackId) throws TrackNotFoundException {
         Optional<Track> userOptional = muzixRepository.findById(trackId);
-        if (!userOptional.isPresent())
-            return track;
+        if (!userOptional.isPresent()) {
+            throw new TrackNotFoundException("id not found");
+
+        }
         track.setTrackid(trackId);
         muzixRepository.save(track);
         return track;
     }
 
     @Override
-
     public List<Track> getByName(String name) {
         List<Track> tracks=muzixRepository.findTitleByName(name);
         return tracks;
     }
+
 }
